requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
Django==5.1.1
django-appconf==1.0.6
django-staticfiles==1.2.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
